tests/wpt/tests/css/css-gaps/tentative/parsing/rule-paint-order-valid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: parsing gap-rule-paint-order valid values</title>
  <link rel="author" title="Javier Contreras Tenorio" href="mailto:javiercon@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#paint-order">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_valid_value('gap-rule-paint-order', 'row-over-column');
    test_valid_value('gap-rule-paint-order', 'column-over-row');
  </script>
</body>
</html>
